Moved link toolbar positioning into link toolbar element
no issue

- keeps related functionality together, cleans up the main floating toolbar component
